deleted image picker file and extracted required logic to Request file
