+++ client/src/utils/Requests.js
@@ -1,6 +1,8 @@
 import React from 'react';
-import { AUTH0_CLIENT, AUTH0_DOMAIN, BASE_URL, LOCAL } from 'react-native-dotenv';
+import { AUTH0_CLIENT, AUTH0_DOMAIN, BASE_URL, LOCAL, AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY } from 'react-native-dotenv';
 import Auth0 from 'react-native-auth0';
+import ImagePicker from 'react-native-image-picker';
+import { RNS3 } from 'react-native-aws3';
 const auth0ClientId = AUTH0_CLIENT;
 const auth0Domain = AUTH0_DOMAIN;
 const local = `http://localhost:3000`;
@@ -53,14 +55,14 @@ export const handleAuth = async (dispatch) => {
   try {
 
     const getAuth = await auth0.webAuth
-    .authorize({
-      scope: 'openid profile email offline_access',
-      audience: 'https://lambdasocial.auth0.com/userinfo',
-      prompt: 'login'
-    })
+      .authorize({
+        scope: 'openid profile email offline_access',
+        audience: 'https://lambdasocial.auth0.com/userinfo',
+        prompt: 'login'
+      })
     const getUserWithAuth = await getUser(getAuth.accessToken, dispatch); // send access_token
     return getUserWithAuth;
-  } catch(error) {
+  } catch (error) {
     console.log('error in login', error);
   }
 };
@@ -90,17 +92,17 @@ const makeUser = async (token, info) => {
     email: info.email,
     avatar: info.picture
   }); // send  nickname as a 'username'
-try {
-  const postUser = await fetch(`${BASE_URL}/users`, {
-    method: 'POST',
-    headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json'
-    },
-    body
-  })
-  return postUser
-} catch(error) {
+  try {
+    const postUser = await fetch(`${BASE_URL}/users`, {
+      method: 'POST',
+      headers: {
+        Accept: 'application/json',
+        'Content-Type': 'application/json'
+      },
+      body
+    })
+    return postUser
+  } catch (error) {
     console.log('error in sending user', error);
   };
 };
@@ -110,4 +112,51 @@ export const handleLogout = async (dispatch) => {
   AsyncStorage.removeItem('accessToken');
   dispatch({ type: 'LOGOUT' });
   return <Redirect to="/home" />;
+}
+
+// handles aws image uploading
+export const uploadImage = () => {
+  ImagePicker.showImagePicker({}, (response) => {
+    /*response returns an object with all of the information about the selected image.
+    returns data, fileName, fileSize, height, isVertical, latitude, longitude, origURL,
+    timestamp, type, uri, width.
+    */
+    // extract this data for the file upload
+    const file = {
+      uri: response.uri,
+      name: response.fileName,
+      type: 'image/png'
+    }
+    console.log(file)
+    // s3 configurations
+    const config = {
+      keyPrefix: 's3/',
+      bucket: 'lambdasocialbucket',
+      region: 'us-east-1',
+      accessKey: AWS_ACCESS_KEY_ID,
+      secretKey: AWS_SECRET_ACCESS_KEY,
+      successActionStatus: 201
+    }
+    RNS3.put(file, config)
+      .then((response) => {
+        if (response.status === 403) {
+          console.log(response);
+          throw new Error("Failed to upload image to S3");
+        } else {
+          /*
+          response will come back looking like this, we'll want
+          the location for the POST request to make a discussion.
+              location: "https://lambdasocialbucket.s3.amazonaws.com/s3%2FIMG_0111.HEIC"
+          }
+              */
+
+          //TODO: remove this console log and replace it with code
+          console.log(response.body.postResponse.location);
+        }
+
+      })
+      .catch((err) => {
+        console.log(err)
+      })
+  })
 }
